refactor: LottoService 로또 결과 값 Controller로 이전
- profit, lottoRankingStatus -> LottoController
